+++ include/common/UpdateChecker.h
@@ -13,7 +13,7 @@
 #include <stdio.h>
 
 /** 当前软件版本号 */
-#define HIS_VERSION "7.2.3"
+#define HIS_VERSION "7.3.0"
 
 /**
  * @brief 更新信息结构体